pyproject.toml
[tool.poetry]
name = "ml-training-api"
version = "0.1.0"
description = "ML Model Training API with REST and gRPC interfaces"
authors = ["emilia, anna, polina"]
packages = [{include = "*.py"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.104.1"
uvicorn = "0.24.0"
scikit-learn = "1.3.2"
pandas = "1.5.3"
numpy = "1.24.3"
grpcio = "1.59.3"
grpcio-tools = "1.59.3"
gradio = "^3.50.0"
plotly = "5.17.0"
requests = "2.31.0"
joblib = "1.3.2"
pydantic = "2.5.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"